40 add authentication to dashboard (#41)
* Added password input

* Added a cookie

* Remove src in package import

* Remove src

* Test local imports

* Rename dashboard
