+++ .gitignore
@@ -171,5 +171,6 @@ secret.key
 # Poetry file.
 poetry.lock
 
-#streamlit stuff
+# Streamlit stuff
+.streamlit/
 secrets.toml
