+++ src/dashboard/plots.py
@@ -1,12 +1,9 @@
-"""dashboard.py - Streamlit app for displaying the stats dashboard.
-"""
+"""plots.py - Create plots for the dashboard"""
 
-import subprocess
-import pymongo
-import streamlit as st
-import pandas as pd
+# Get packages.
 import altair as alt
-from log_keeper.get_config import Config
+import pandas as pd
+import streamlit as st
 
 
 def format_data_for_table(raw_df: pd.DataFrame) -> pd.DataFrame:
@@ -58,38 +55,6 @@ def format_data_for_table(raw_df: pd.DataFrame) -> pd.DataFrame:
     return data_df
 
 
-def fetch_data_from_mongodb() -> pd.DataFrame:
-    """Fetch data from MongoDB and return as a DataFrame.
-    Returns:
-        pd.DataFrame: The data from MongoDB."""
-
-    try:
-        # Construct the MongoDB connection URI
-        db_config = Config()
-        if not db_config.validate():
-            db_config.update_credentials()
-
-        # Create the DB connection URL
-        db_url = (
-            f"mongodb+srv://{db_config.db_username}:"
-            f"{db_config.db_password}@{db_config.db_hostname}"
-            "/?retryWrites=true&w=majority"
-        )
-
-        # Connect to MongoDB
-        client = pymongo.MongoClient(db_url)
-        db = client[db_config.db_name]
-        collection = db[db_config.db_collection_name]
-
-        # Convert list of dictionaries to DataFrame
-        df = pd.DataFrame(collection.find())
-
-    except Exception as e:  # pylint: disable=broad-except
-        st.error(f"Error: {e}")
-        df = pd.DataFrame()
-    return df
-
-
 def plot_launches_by_commander(df: pd.DataFrame):
     """ Plot the number of launches by AircraftCommander.
 
@@ -119,6 +84,35 @@ def plot_launches_by_commander(df: pd.DataFrame):
 
     # Display the chart in Streamlit.
     st.altair_chart(chart, use_container_width=True)
+
+
+def plot_all_launches(df: pd.DataFrame):
+    """ Plot all launches in the data.
+
+    Args:
+        df (pd.DataFrame): The data to be plotted.
+    """
+    # Sort the data by date in descending order.
+    df = df.sort_values(by="Date", ascending=False)
+    df = df.drop(columns=["_id"])
+
+    # Format the date.
+    df["Date"] = df["Date"].dt.strftime("%d %b %y")
+
+    # Convert the FlightTime (minutes) to a string in HH:MM format.
+    df["FlightTime"] = df["FlightTime"].apply(
+        lambda x: f"{x//60}:{x % 60:02d}"
+    )
+
+    # Format TakeOffTime and LandingTime.
+    df["TakeOffTime"] = df["TakeOffTime"].dt.strftime("%H:%M")
+    df["LandingTime"] = df["LandingTime"].dt.strftime("%H:%M")
+
+    # Make date the first column.
+    df = df[["Date"] + [col for col in df.columns if col != "Date"]]
+
+    # Plot all data.
+    st.dataframe(df, use_container_width=True)
 
 
 def show_logbook_helper(df: pd.DataFrame, commander: str):
@@ -158,54 +152,6 @@ def show_logbook_helper(df: pd.DataFrame, commander: str):
         hide_index=True,
         use_container_width=True
     )
-
-
-def date_filter(df: pd.DataFrame) -> pd.DataFrame:
-    """Filter the data by date.
-
-    Args:
-        df (pd.DataFrame): The data to be filtered.
-
-    Returns:
-        pd.DataFrame: The filtered data.
-    """
-    # Add a date filter to the sidebar.
-    st.sidebar.markdown("## Date Filter")
-
-    # Get the date range from the user.
-    min_date = df["Date"].min()
-    max_date = df["Date"].max()
-
-    # Add a date filter to the sidebar.
-    start_date = st.sidebar.date_input(
-        "Start Date",
-        value=min_date,
-        min_value=min_date,
-        max_value=max_date,
-        help="Select the start date",
-    )
-    end_date = st.sidebar.date_input(
-        "End Date",
-        value=max_date,
-        min_value=min_date,
-        max_value=max_date,
-        help="Select the end date",
-    )
-
-    # Convert the date to a pandas datetime object.
-    start_date = pd.to_datetime(start_date)
-    end_date = pd.to_datetime(end_date)
-
-    # Validate the date range.
-    if start_date <= end_date:
-        # Filter the data by the date range.
-        filtered_df = df[
-            (df["Date"] >= start_date) & (df["Date"] <= end_date)
-        ]
-    else:
-        st.error("Error: End date must fall after start date.")
-        filtered_df = df
-    return filtered_df
 
 
 def quarterly_summary(df: pd.DataFrame,
@@ -275,105 +221,3 @@ def quarterly_summary(df: pd.DataFrame,
         data=summary,
         hide_index=True,
     )
-
-
-def plot_all_launches(df: pd.DataFrame):
-    """ Plot all launches in the data.
-
-    Args:
-        df (pd.DataFrame): The data to be plotted.
-    """
-    # Sort the data by date in descending order.
-    df = df.sort_values(by="Date", ascending=False)
-    df = df.drop(columns=["_id"])
-
-    # Format the date.
-    df["Date"] = df["Date"].dt.strftime("%d %b %y")
-
-    # Convert the FlightTime (minutes) to a string in HH:MM format.
-    df["FlightTime"] = df["FlightTime"].apply(
-        lambda x: f"{x//60}:{x % 60:02d}"
-    )
-
-    # Format TakeOffTime and LandingTime.
-    df["TakeOffTime"] = df["TakeOffTime"].dt.strftime("%H:%M")
-    df["LandingTime"] = df["LandingTime"].dt.strftime("%H:%M")
-
-    # Make date the first column.
-    df = df[["Date"] + [col for col in df.columns if col != "Date"]]
-
-    # Plot all data.
-    st.dataframe(df, use_container_width=True)
-
-
-def main():
-    """Main Streamlit App Code."""
-    # Set the page title.
-    st.markdown("# 661 VGS Dashboard")
-    st.sidebar.markdown("# Dashboard Filters")
-
-    # Fetch data from MongoDB
-    if "df" not in st.session_state:
-        st.session_state.df = fetch_data_from_mongodb()
-
-    # Refresh data button.
-    if st.button("🔃 Refresh Data"):
-        st.session_state.df = fetch_data_from_mongodb()
-        st.success("Data Refreshed!", icon="✅")
-
-    # Get the data from the session state.
-    df = st.session_state.df
-
-    # Filter by AircraftCommander.
-    commander = st.sidebar.selectbox(
-        "Filter by AircraftCommander",
-        sorted(df["AircraftCommander"].unique()),
-        index=None,
-        help="Select the AircraftCommander to filter by.",
-        placeholder="All",
-    )
-
-    # Add a date filter to the sidebar.
-    filtered_df = date_filter(df)
-
-    tabs = ["Statistics", "All Data"]
-    stats_tab, all_data_tab = st.tabs(tabs)
-
-    with stats_tab:
-
-        # Plot the number of launches by unique AircraftCommander.
-        plot_launches_by_commander(filtered_df)
-
-        # Logbook helper by AircraftCommander.
-        show_logbook_helper(filtered_df, commander)
-
-        # Show a quarterly summary of the number of launches
-        # for each AircraftCommander.
-        st.sidebar.markdown("## Quarterly Summary")
-
-        # Create a list of quarters from the data.
-        quarters = filtered_df["Date"].dt.to_period("Q").unique()
-
-        quarter = st.sidebar.selectbox(
-            "Select Quarter",
-            quarters,
-            index=None,
-            help="Select the quarter to display."
-        )
-
-        # Filter the data by the selected quarter.
-        if quarter and commander:
-            quarterly_summary(filtered_df, commander, quarter)
-
-    with all_data_tab:
-        plot_all_launches(filtered_df)
-
-
-def display_dashboard():
-    """Run the Streamlit app."""
-    subprocess.run(["streamlit", "run", "src/dashboard/dashboard.py"],
-                   check=True)
-
-
-if __name__ == '__main__':
-    main()
